01AICodeSnippets/06-multimodal-ai/requirements.txt
# Core dependencies
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
python-dotenv>=0.19.0
pydantic>=1.8.2

# Vision processing
google-generativeai>=0.3.0
Pillow>=9.0.0

# Audio processing
openai-whisper>=20231117
torch>=2.0.0
numpy>=1.21.0

# Utilities
python-jose>=3.3.0
requests>=2.26.0 